+++ Source/PlaneGameObject.cpp
@@ -82,7 +82,7 @@ void PlaneGameObject::InitMeshDataAndBuffers(DirectX::SimpleMath::Vector2Int pla
 	D3D11_BUFFER_DESC bd;
 	ZeroMemory(&bd, sizeof(bd));
 	bd.Usage = D3D11_USAGE_DEFAULT;
-	bd.ByteWidth = sizeof(Vertex) * vertices.size();
+	bd.ByteWidth = static_cast<UINT>(sizeof(Vertex) * vertices.size());
 	bd.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 	bd.CPUAccessFlags = 0;
 
@@ -98,7 +98,7 @@ void PlaneGameObject::InitMeshDataAndBuffers(DirectX::SimpleMath::Vector2Int pla
 	D3D11_BUFFER_DESC iBd;
 	ZeroMemory(&iBd, sizeof(iBd));
 	iBd.Usage = D3D11_USAGE_DEFAULT;
-	iBd.ByteWidth = sizeof(unsigned short) * indices.size();
+	iBd.ByteWidth = static_cast<UINT>(sizeof(unsigned short) * indices.size());
 	iBd.BindFlags = D3D11_BIND_INDEX_BUFFER;
 	iBd.CPUAccessFlags = 0;
 
@@ -110,7 +110,7 @@ void PlaneGameObject::InitMeshDataAndBuffers(DirectX::SimpleMath::Vector2Int pla
 										   &sIbd,
 										   s_planeIndexBuffer.ReleaseAndGetAddressOf()));
 	
-	s_indexCount = indices.size();
+	s_indexCount = (int)indices.size();
 
 	delete[] s_planeVertices;
 	delete[] s_planeIndices;
@@ -128,9 +128,9 @@ void PlaneGameObject::InitDebugTexture(const wchar_t* texturePath, ID3D11Device*
 std::vector<Vertex> PlaneGameObject::GeneratePlaneVertices(int width, int depth) {
 	std::vector<Vertex> returnVec;
 
-	for (unsigned int y = 0; y < depth; y++) {
-		for (unsigned int x = 0; x < width; x++) {
-			Vertex vert = { XMFLOAT3(x, 0, y), XMFLOAT3(0, 1, 0), XMFLOAT2(x, y) };
+	for (int y = 0; y < depth; y++) {
+		for (int x = 0; x < width; x++) {
+			Vertex vert = { XMFLOAT3((float)x, 0.f, (float)y), XMFLOAT3(0.f, 1.f, 0.f), XMFLOAT2((float)x, (float)y) };
 			returnVec.push_back(vert);
 		}
 	}
@@ -141,17 +141,17 @@ std::vector<Vertex> PlaneGameObject::GeneratePlaneVertices(int width, int depth)
 std::vector<unsigned short> PlaneGameObject::GeneratePlaneIndices(int width, int depth) {
 	std::vector<unsigned short> returnVec;
 
-	for (unsigned int y = 0; y < depth - 1; y++) {
-		for (unsigned int x = 0; x < width - 1; x++) {
+	for (int y = 0; y < depth - 1; y++) {
+		for (int x = 0; x < width - 1; x++) {
 			//Triangle 1
-			returnVec.push_back((y)*width + (x));
-			returnVec.push_back((y + 1) * width + (x));
-			returnVec.push_back((y)*width + (x + 1));
+			returnVec.push_back(static_cast<unsigned short>((y) * width + (x)));
+			returnVec.push_back(static_cast<unsigned short>((y + 1) * width + (x)));
+			returnVec.push_back(static_cast<unsigned short>((y)*width + (x + 1)));
 
 			//Triangle 2
-			returnVec.push_back((y)*width + (x + 1));
-			returnVec.push_back((y + 1) * width + (x));
-			returnVec.push_back((y + 1) * width + (x + 1));
+			returnVec.push_back(static_cast<unsigned short>((y) * width + (x + 1)));
+			returnVec.push_back(static_cast<unsigned short>((y + 1) * width + (x)));
+			returnVec.push_back(static_cast<unsigned short>((y + 1) * width + (x + 1)));
 		}
 	}
